BREAKING: move awful powerdns treatment under terraform provider
